+++ R/gpkg-util.R
@@ -50,7 +50,7 @@ gpkg_source.geopackage <- function(x) {
 
 #' List Tables in a GeoPackage
 #'
-#' @param x A _geopackage_ object, path to a GeoPackage or an `SQLiteConnection`
+#' @param x A _geopackage_ object, path to a GeoPackage or an _SQLiteConnection_
 #' @return a character vector with names of all tables and views in the database
 #' @importFrom DBI dbListTables dbDisconnect
 #' @export
@@ -68,7 +68,7 @@ gpkg_list_tables <- function(x) {
 
 #' Set `data_null` Metadata for a GeoPackage Tile Dataset
 #'
-#' @param x A _geopackage_ object, path to a GeoPackage or an `SQLiteConnection`
+#' @param x A _geopackage_ object, path to a GeoPackage or an _SQLiteConnection_
 #' @param name character. Tile matrix set name(s) (`tile_matrix_set_name`)
 #' @param value numeric. Value to use as "NoData" (`data_null` value)
 #' @param query_string logical. Return SQLite query rather than executing it? Default: `FALSE`
@@ -94,7 +94,7 @@ gpkg_tile_set_data_null <- function(x, name, value, query_string = FALSE) {
 
 #' Get `gpkg_2d_gridded_coverage_ancillary` Table
 #'
-#' @param x A _geopackage_ object, path to a GeoPackage or an `SQLiteConnection`
+#' @param x A _geopackage_ object, path to a GeoPackage or an _SQLiteConnection_
 #' @return a data.frame containing columns `id`, `tile_matrix_set_name`, `datatype`, `scale`, `offset`, `precision`, `data_null`, `grid_cell_encoding`, `uom`, `field_name`, `quantity_definition`
 #' @importFrom DBI dbDisconnect
 #' @export
@@ -103,17 +103,4 @@ gpkg_2d_gridded_coverage_ancillary <- function(x) {
     stop('package `RSQLite` is required to get the `gpkg_2d_gridded_coverage_ancillary` table', call. = FALSE)
   }
   gpkg_table(x, "gpkg_2d_gridded_coverage_ancillary", collect = TRUE)
-}
-
-#' Get `gpkg_contents` Table
-#'
-#' @param x A _geopackage_ object, path to a GeoPackage or an `SQLiteConnection`
-#' @return a data.frame containing columns `table_name`, `data_type`, `identifier`, `description`, `last_change`, `min_x`, `min_y`, `max_x`, `max_y`, `srs_id`
-#' @importFrom DBI dbDisconnect
-#' @export
-gpkg_contents <- function(x) {
-  if (!requireNamespace("RSQLite", quietly = TRUE)) {
-    stop('package `RSQLite` is required to get the `gpkg_contents` table', call. = FALSE)
-  }
-  gpkg_collect(x, "gpkg_contents")
 }